rename CNI from ptpnat to netglue
this better reflects the current use-cases
